first commit
Undocumented code. Basically a copy from 'simple game' published on libgdx's wiki with the addition of a random number generator that produces different sounds and the use of different images which includes a background image. Images are not set to size of object, causing lack of cleanliness in game design (trump heads have to hit certain spot in cage).